CV_project/Model_2:SiameseNetwork/requirements.txt
pytorch
numpy
scipy
sklearn
pillow
